pyproject.toml
[tool.poetry]
name = "larixon-test"
version = "0.1.0"
description = ""
authors = ["nurj0hn <nurjantunteev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
celery = "^5.3.4"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
django-debug-toolbar = "^4.2.0"
coreapi = "^2.3.3"
django-storages = "^1.14.1"
django-cors-headers = "^4.2.0"
boto3 = "^1.28.59"
django-celery-results = "^2.5.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
